feat(place): support per-category fetch and detail lookup in TourApiClient
fetchRestaurantsNear(고정 contentTypeId=39)를 fetchPlacesNear(category, ...)로
바꿔 카테고리별 contentTypeId/cat3 파라미터를 동적으로 구성하도록 했다.
detailCommon2 + detailIntro2를 합쳐 raw JSON으로 반환하는 fetchDetail()을
추가했다. 실제 서비스키로 호출해 확인한 결과를 반영해 다음을 수정했다:
- KTO 응답은 header/body가 아니라 response.header/response.body로 감싸져
  있고, 파라미터 오류 응답은 반대로 래퍼 없이 resultCode가 최상위에 온다
  (extractBody()로 두 형태를 모두 처리).
- detailCommon2는 contentTypeId 파라미터를 받으면
  INVALID_REQUEST_PARAMETER_ERROR를 반환하므로 contentId만 전달한다.
